refactor: rename cached to cached_repository
